A11yPuppetry report UI (WIP) (#30)
* Annotate acted elements on replay report

* Use TalkBack Directional Gif

* Add results to background of actions

* Change Table View to Column View in Replay Report

* Add background colors & notes

+++ py_src/data_utils.py
@@ -1,21 +1,23 @@
 import json
 import shutil
+from collections import defaultdict
 from typing import List
+from cachetools import cached, TTLCache
 
 from app import App
-from command import Command, create_command_response_from_dict, create_command_from_dict
+from command import Command, create_command_response_from_dict, create_command_from_dict, LocatableCommandResponse
 from consts import BLIND_MONKEY_EVENTS_TAG
 from snapshot import Snapshot
 
 
 class ReplayDataManager:
-    def __init__(self, app: App, controller_mode: str):
+    def __init__(self, app: App, controller_mode: str, recreate: bool = False):
         self.app = app
         self.controller_mode = controller_mode
         self.replay_path = app.app_path.joinpath(f"REPLAY_{controller_mode}")
-        if self.replay_path.exists():
+        if self.replay_path.exists() and recreate:
             shutil.rmtree(self.replay_path)
-        self.replay_path.mkdir(parents=True, exist_ok=False)
+        self.replay_path.mkdir(parents=True, exist_ok=True)
         self.replay_usecase_report_path = self.replay_path.joinpath("usecase_report.jsonl")
         self.replay_usecase_report_path.touch()
         self.replay_usecase_finish_path = self.replay_path.joinpath("finish.txt")
@@ -51,6 +53,20 @@ class ReplayDataManager:
                 snapshots.append(self.app.get_snapshot(snapshot_info['snapshot_name']))
         return snapshots
 
+    @cached(cache=TTLCache(maxsize=1024, ttl=10))
+    def get_problematic_steps(self) -> dict:
+        problems = defaultdict(list)
+        for snapshot in self.get_snapshots():
+            snapshot_index = snapshot.name.split("_")[-1]
+
+            step_info = self.get_step_info(snapshot_index)
+            if step_info['response'].state != 'COMPLETED':
+                reason = 'Unknown'
+                reason = step_info['response'].state
+                problems[snapshot_index].append(reason)
+        return problems
+
+    @cached(cache=TTLCache(maxsize=1024, ttl=10))
     def get_step_info(self, index: str) -> dict:
         snapshot = self.app.get_snapshot(name=f"{self.controller_mode}.S_{index}")
         step_info = {
@@ -68,14 +84,24 @@ class ReplayDataManager:
                 step_info_json = json.load(f)
 
             step_info['command'] = create_command_from_dict(step_info_json.get('command', {}))
-            step_info['response'] = create_command_response_from_dict(step_info['command'],
+            response = step_info['response'] = create_command_response_from_dict(step_info['command'],
                                                                   step_info_json.get('response', {}))
+            if isinstance(response, LocatableCommandResponse):
+                screen_bounds = snapshot.nodes[0].bounds  # TODO: Not correct when the keyboard is enabled
+                if screen_bounds[0] != 0:
+                    screen_bounds = [0, 0, 1080, 2340]  # TODO: Move to consts
+                step_info['bounds'] = str(list(response.acted_node.get_normalized_bounds(screen_bounds)))
+            else:
+                step_info['bounds'] = "[0.0,0.0,0.0,0.0]"
             step_info['logs'] = snapshot.address_book.get_log_path(mode=self.controller_mode, index=0)
             step_info['event_logs'] = snapshot.address_book.get_log_path(mode=self.controller_mode, index=0,
                                                                      extension=BLIND_MONKEY_EVENTS_TAG)
 
         step_info['layout'] = snapshot.address_book.get_layout_path(mode=self.controller_mode, index=0)
-        step_info['screenshot'] = snapshot.initial_screenshot
+        if self.controller_mode == 'tb_dir' and snapshot.address_book.tb_explore_visited_nodes_gif.exists():
+            step_info['screenshot'] = snapshot.address_book.tb_explore_visited_nodes_gif
+        else:
+            step_info['screenshot'] = snapshot.initial_screenshot
 
         return step_info
